Enable option to ignore output EOS during training
- When retrieving the target during training, eos symbols are replaced by pad symbols.
- Makes sure that Evaluator can work without SupervisedTrainer. get_batch_data is now implemented in both trainer and evaluator.
Instead of also creating an attention_evaluator just as the attention_trainer is removed the attention_trainer class and just passed the attention_function to the train() and evaluate() function.
It is not best practice to have identical get_batch_data function in both trainer, evaluator (and predictor). We might want to move this functionality to a different class.