Add new method for attention integration
Instead of simply merging the context vector and the embedded input for the decoder,
the option 'focus_attention' in the decoder allows more flexibility in how
the two are combined. In the current implementation, the context vector and the
embeddings are first concatenated; then an affine transformation is applied
to project them in the same space; then RELU is applied; then, and this is
the key of this new method, an elementwise multiplication between the
combined inputs and the context vector is applied.